app/requirements.txt
pyserial >= 3.5
PyQt6 >= 6.4.2
pyqtgraph >= 0.13.1
